requirements.txt
inferedge-moss>=0.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
openai>=1.47.0
transformers>=4.44.0
optimum[onnxruntime]>=1.21.0
torch>=2.4.0
sentence-transformers>=3.1.0
numpy>=1.26.4
tqdm>=4.66.5
orjson>=3.10.7
pytest>=8.3.3
